requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
-e git+https://github.com/City-of-Helsinki/munigeo@v0.2.3#egg=django-munigeo
-e git+https://github.com/City-of-Helsinki/django-orghierarchy.git@v0.1.4#egg=django-orghierarchy
astroid==1.6.1            # via pylint
attrs==17.4.0
bleach==2.1.3
certifi==2018.1.18        # via requests
chardet==3.0.4            # via requests
click==6.7                # via pip-tools
defusedxml==0.5.0         # via python3-openid
django-allauth==0.35.0
django-appconf==1.0.1     # via django-image-cropping
django-cleanup==2.1.0
django-cors-headers==2.2.0
django-extensions==2.0.3
django-filter==1.1.0
django-haystack==2.7.0
django-helusers==0.4.2
django-image-cropping==1.1.0
django-js-asset==1.0.0    # via django-mptt
django-leaflet==0.23.0
django-modeltranslation==0.12.2
django-mptt==0.8.7
django-reversion==2.0.13
django==1.11.15
djangorestframework-bulk==0.2.1
djangorestframework-gis==0.12
djangorestframework-jwt==1.11.0
djangorestframework==3.7.7
drf-oidc-auth==0.9        # via django-helusers
easy-thumbnails==2.5
elasticsearch==2.4.1
first==2.0.1              # via pip-tools
flake8==3.5.0
future==0.16.0            # via pyjwkest
html5lib==1.0.1           # via bleach
httmock==1.2.6
icalendar==4.0.1
idna==2.6                 # via requests
isodate==0.6.0
isort==4.3.4              # via pylint
lazy-object-proxy==1.3.1  # via astroid
lxml==4.1.1
markdown==2.6.11
mccabe==0.6.1             # via flake8, pylint
oauthlib==2.0.6           # via requests-oauthlib
olefile==0.44             # via pillow
pillow==5.0.0
pip-tools==1.11.0
pluggy==0.6.0             # via pytest
psycopg2==2.7.4
py==1.5.2                 # via pytest
pycodestyle==2.3.1        # via flake8
pycryptodomex==3.5.1      # via pyjwkest
pyflakes==1.6.0
pyjwkest==1.4.0           # via drf-oidc-auth
pyjwt==1.6.0              # via djangorestframework-jwt
pylint==1.8.2
pyparsing==2.2.0          # via rdflib
pytest-django==3.1.2
pytest==3.4.2             # via pytest-django
python-dateutil==2.6.1
python-docx==0.8.6
python3-openid==3.1.0     # via django-allauth
pytz==2018.3
pyyaml==3.12
raven==6.6.0
rdflib==4.2.2
requests-cache==0.4.13
requests-oauthlib==0.8.0  # via django-allauth
requests==2.20.0
six==1.11.0
swapper==1.1.0
urllib3==1.22             # via elasticsearch, requests
webencodings==0.5.1       # via html5lib
wrapt==1.10.11            # via astroid
django-extra-fields==1.0.0  # Tavastia Events
azure==1.0.3              # Tavastia Events
django-storages==1.6.6    # Tavastia Events